*: make storeTestSuite acceleration
