+++ tests/server/api/region_test.go
@@ -61,54 +61,9 @@ func (suite *regionTestSuite) TearDownSuite() {
 }
 
 func (suite *regionTestSuite) TearDownTest() {
-	cleanFunc := func(cluster *tests.TestCluster) {
-		re := suite.Require()
-		// clean rules
-		defaultRule := placement.GroupBundle{
-			ID: "pd",
-			Rules: []*placement.Rule{
-				{GroupID: "pd", ID: "default", Role: "voter", Count: 3},
-			},
-		}
-		data, err := json.Marshal([]placement.GroupBundle{defaultRule})
-		re.NoError(err)
-		urlPrefix := cluster.GetLeaderServer().GetAddr()
-		err = testutil.CheckPostJSON(tests.TestDialClient, urlPrefix+"/pd/api/v1/config/placement-rule", data, testutil.StatusOK(re))
-		re.NoError(err)
-		// clean region cache
-		for _, server := range cluster.GetServers() {
-			pdAddr := cluster.GetConfig().GetClientURL()
-			for _, region := range server.GetRegions() {
-				url := fmt.Sprintf("%s/pd/api/v1/admin/cache/region/%d", pdAddr, region.GetID())
-				err := testutil.CheckDelete(tests.TestDialClient, url, testutil.StatusOK(re))
-				re.NoError(err)
-			}
-			re.Empty(server.GetRegions())
-		}
-
-		// clean stores
-		leader := cluster.GetLeaderServer()
-		for _, store := range leader.GetStores() {
-			if store.NodeState == metapb.NodeState_Removed {
-				continue
-			}
-			re.NoError(cluster.GetLeaderServer().GetRaftCluster().RemoveStore(store.GetId(), true))
-			re.NoError(cluster.GetLeaderServer().GetRaftCluster().BuryStore(store.GetId(), true))
-		}
-		re.NoError(cluster.GetLeaderServer().GetRaftCluster().RemoveTombStoneRecords())
-		re.Empty(leader.GetStores())
-		testutil.Eventually(re, func() bool {
-			if sche := cluster.GetSchedulingPrimaryServer(); sche != nil {
-				for _, s := range sche.GetBasicCluster().GetStores() {
-					if s.GetState() != metapb.StoreState_Tombstone {
-						return false
-					}
-				}
-			}
-			return true
-		})
-	}
-	suite.env.RunTest(cleanFunc)
+	re := suite.Require()
+	suite.env.RunTest(cleanRules(re))
+	suite.env.RunTest(cleanStoresAndRegions(re))
 }
 
 func (suite *regionTestSuite) TestAccelerateRegionsScheduleInRange() {
@@ -344,20 +299,6 @@ func checkRegionCount(re *require.Assertions, cluster *tests.TestCluster, count 
 		testutil.Eventually(re, func() bool {
 			return sche.GetCluster().GetRegionCount([]byte{}, []byte{}) == int(count)
 		})
-	}
-}
-
-func pauseAllCheckers(re *require.Assertions, cluster *tests.TestCluster) {
-	checkerNames := []string{"learner", "replica", "rule", "split", "merge", "joint-state"}
-	addr := cluster.GetLeaderServer().GetAddr()
-	for _, checkerName := range checkerNames {
-		resp := make(map[string]any)
-		url := fmt.Sprintf("%s/pd/api/v1/checker/%s", addr, checkerName)
-		err := testutil.CheckPostJSON(tests.TestDialClient, url, []byte(`{"delay":1000}`), testutil.StatusOK(re))
-		re.NoError(err)
-		err = testutil.ReadGetJSON(re, tests.TestDialClient, url, &resp)
-		re.NoError(err)
-		re.True(resp["paused"].(bool))
 	}
 }
