+++ tests/server/api/store_test.go
@@ -48,12 +48,17 @@ func TestStoreTestSuite(t *testing.T) {
 	suite.Run(t, new(storeTestSuite))
 }
 
-func (suite *storeTestSuite) SetupTest() {
+func (suite *storeTestSuite) SetupSuite() {
 	suite.env = tests.NewSchedulingTestEnvironment(suite.T())
 }
 
-func (suite *storeTestSuite) TearDownTest() {
+func (suite *storeTestSuite) TearDownSuite() {
 	suite.env.Cleanup()
+}
+func (suite *storeTestSuite) TearDownTest() {
+	re := suite.Require()
+	suite.env.RunTest(cleanRules(re))
+	suite.env.RunTest(cleanStoresAndRegions(re))
 }
 
 func (suite *storeTestSuite) TestStoresList() {
